% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier_option_price.R
\name{barrier_option_price}
\alias{barrier_option_price}
\title{European-style up-and-out}
\usage{
barrier_option_price(nInt, strike, spot, vol, r, expiry, barrier, nReps)
}
\arguments{
\item{nInt}{Number of time intervals}

\item{strike}{Strike price}

\item{spot}{Spot price}

\item{vol}{Volatility}

\item{r}{Risk-free interest}

\item{expiry}{Time to maturity}

\item{barrier}{Barrier level}

\item{nReps}{Number of Monte Carlo sim}
}
\value{
The theoretical price
}
\description{
European-style up-and-out
}
\examples{
price <- barrier_option_price(126, 110, 105, 0.21, 0.05, 0.75, 120, 10000)
print(price)
}
